feat: flujo integral de moderacion de administrador, advertencias oficiales y reportes de usuarios
- Visualizacion en tiempo real de llamadas de atencion oficiales (strikes) en paneles de vendedor y comprador.

- Componente reutilizable ReportIncidentDialog con motivos adaptados para perfiles, seguimiento de pedido y resumen poscompra.

- Centro de Moderacion y Reportes en AdminHomeScreen con gestion de incidencias, resolucion y sancion directa.

- Pantalla completa de detalle de comprador (AdminBuyerDetailScreen) y directorio especializado de estudiantes.

- Actualizacion de politicas RLS y funcion RPC resolve_incident_rpc en Supabase.

+++ app/src/main/java/com/example/vallego/domain/model/CampusDetailedMetrics.kt
@@ -18,7 +18,7 @@ data class SellerSalesRanking(
     @SerialName("avatar_url") val avatarUrl: String? = null,
     @SerialName("total_sales") val totalSales: Double = 0.0,
     @SerialName("completed_orders") val completedOrders: Int = 0,
-    @SerialName("percentage") val percentage: Float = 0.0f
+    @SerialName("percentage") val percentage: Double = 0.0
 )
 
 @Serializable
@@ -26,14 +26,14 @@ data class PaymentMethodBreakdown(
     @SerialName("method") val method: String,
     @SerialName("count") val count: Int = 0,
     @SerialName("total_amount") val totalAmount: Double = 0.0,
-    @SerialName("percentage") val percentage: Float = 0.0f
+    @SerialName("percentage") val percentage: Double = 0.0
 )
 
 @Serializable
 data class MeetingPointTraffic(
     @SerialName("point_name") val pointName: String,
     @SerialName("count") val count: Int = 0,
-    @SerialName("percentage") val percentage: Float = 0.0f
+    @SerialName("percentage") val percentage: Double = 0.0
 )
 
 @Serializable
@@ -44,7 +44,7 @@ data class CampusDetailedMetrics(
     @SerialName("completed_orders") val completedOrders: Int = 0,
     @SerialName("cancelled_orders") val cancelledOrders: Int = 0,
     @SerialName("average_ticket") val averageTicket: Double = 0.0,
-    @SerialName("fulfillment_rate") val fulfillmentRate: Float = 100.0f,
+    @SerialName("fulfillment_rate") val fulfillmentRate: Double = 100.0,
     @SerialName("seller_rankings") val sellerRankings: List<SellerSalesRanking> = emptyList(),
     @SerialName("payment_methods") val paymentMethods: List<PaymentMethodBreakdown> = emptyList(),
     @SerialName("top_meeting_points") val topMeetingPoints: List<MeetingPointTraffic> = emptyList()
